% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_nlmixr_model.R
\name{summarise_nlmixr2_model}
\alias{summarise_nlmixr2_model}
\title{Data summary function}
\usage{
summarise_nlmixr2_model(obj, model, software, rounding, runname)
}
\arguments{
\item{obj}{nlmixr2 fit object to be evaluated}

\item{model}{Model. Can be blank}

\item{software}{Software that generated the model fit}

\item{rounding}{Number of figures to round estimates to}

\item{runname}{Name of the model object being converted}
}
\value{
A summary data object used by \link{xpose_data_nlmixr2}.
}
\description{
Convert 'nlmixr2' model output into an 'xpose' database
}
