% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_xp_nlmixr.R
\name{theme_xp_nlmixr2}
\alias{theme_xp_nlmixr2}
\title{Default 'nlmixr2' theme for 'xpose'}
\usage{
theme_xp_nlmixr2()
}
\value{
A list with 'xpose' theme specifiers.
}
\description{
Default 'nlmixr2' theme for 'xpose'.
}
