% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nlmixr_vpc_theme.R
\docType{data}
\name{nlmixr2_vpc_theme}
\alias{nlmixr2_vpc_theme}
\title{Default VPC theme for 'xpose.nlmixr2'}
\format{
An object of class \code{vpc_theme} of length 23.
}
\usage{
nlmixr2_vpc_theme
}
\value{
A list with 'vpc' theme specifiers.
}
\description{
Default VPC theme for 'xpose.nlmixr2'.
}
\keyword{datasets}
